Extracted languages enum to the Language.kt file
